requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3
WTForms==3.1.2
SQLAlchemy==2.0.30
python-dotenv==1.0.1
Markdown==3.6
